~Finish collecting m8[ns] tests, start collecting division by zero tests (#22153)
* implement box fixture, move a couple tests from timedelta.test_arithmetic, parametrize more

* port floordiv tests

* port test_td64arr_rfloordiv_tdlike_scalar

* make fixtures, port last of TimedeltaIndex div/mul tests

* port add/sub tests, cleanup imports

* port test_tdi_add_offset-index

* port offset addsub tests

* start collecting division by zero tests

* dummy commit to force CI

* fix dumb name override mistake
